% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/book_open.R
\name{book_open}
\alias{book_open}
\title{Bring yearly BS book data from sejong}
\usage{
book_open(code)
}
\arguments{
\item{code}{a ticker code in the format as "000000" (use find_code())}
}
\value{
a data frame
}
\description{
\code{book_open} returns major accounting book data
}
\examples{
find_code("stockname") \%>\% book_open
}
