% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_code2.R
\name{find_code2}
\alias{find_code2}
\title{Find ticker code}
\usage{
find_code2(x)
}
\arguments{
\item{x}{is a name of the item to find the ticker code.}
}
\value{
character string of specific ticker code.
}
\description{
\code{find_code} returns ticker code from Daum finance when stock name is given.
}
\examples{
find_code2("stockname")
}
