% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kosd_stock.R
\name{kosd_stock}
\alias{kosd_stock}
\title{Get price history of KOSDAQ stock}
\usage{
kosd_stock(ticker, start_date, end_date)
}
\arguments{
\item{ticker}{a ticker code in the format as KOSDAQ::000000}

\item{start_date}{should be given as a format "2011-01-01"}

\item{end_date}{should be given as a format "2012-01-01"}
}
\value{
a df in the format as OHLC
}
\description{
\code{kosd_stock} returns price history of a given KOSDAQ stock from Google finance
}
\details{
Original: July 2017 by "personalinvesting (naver blog)"
}
\examples{
kosd_stock("011000", "2014-01-01", "2017-09-22")
}
