% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kosp_stock.R
\name{kosp_stock}
\alias{kosp_stock}
\title{Get price history of KOSPI stock}
\usage{
kosp_stock(code, from)
}
\arguments{
\item{ticker}{a ticker code in the format as KRX:000000}

\item{start_date}{should be given as a format "2011-01-01"}

\item{end_date}{should be given as a format "2012-01-01"}
}
\value{
a df in the format as OHLC
}
\description{
\code{kosp_stock} returns price history of a given KOSPI stock from Google finance
}
\examples{
kosp_stock("011000", "2011-01-01")
}
