% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sejong_data.R
\name{sejong_data}
\alias{sejong_data}
\title{Get operating profit data in sejong}
\usage{
sejong_data(ticker, which_item)
}
\arguments{
\item{ticker}{a ticker code in the format as 000000}

\item{which_item}{choose among c("sales", "bp", "np", "asset", "liability", "capital")}
}
\value{
a data frame
}
\description{
Get operating profit data in sejong
}
\examples{
"stock_name" \%>\% find_code2 \%>\% sejong_data("name of statement item")
}
