% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trade_trend.R
\name{trade_trend}
\alias{trade_trend}
\title{Buying and Sell trend of foreign and organization}
\usage{
trade_trend(code, from)
}
\arguments{
\item{code}{a ticker code in the format as "000000" (use find_code())}

\item{from}{should be given as a format "2011-01-01"}
}
\value{
a data frame
}
\description{
\code{trade_trend} returns data frame of trading trends
}
\examples{
find_code("stockname") \%>\% trade_trend("2015-01-01")
}
